refactor(*): move remaining methods in utility to own modules
(cherry picked from commit 7cba5b1c93c8bf298b204a03c822ae7fd09ca6bb)
