+++ webrob/pages/experiments.py
@@ -15,7 +15,7 @@ import StringIO
 from ftplib import FTP
 
 from webrob.app_and_db import app
-from webrob.utility.utility import admin_required
+from webrob.utility.admin_checker import admin_required
 from webrob.models.experiments import Project, Tag
 from webrob.models.db import *
